Day 23-24/requirements.txt
fastapi
pydantic
uvicorn
scikit-learn
django
djangorestframework
djangorestframework-simplejwt
requests